refactor: Move about_store screen into store_details
