+++ lib/ui/screens/store_details/about_store/models/ContactInfo.dart
