+++ lib/ui/screens/store_details/about_store/widgets/address_item.dart
@@ -2,7 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:sellio_mobile/core/design_system/themes/sellio_theme_provider.dart';
 
-import '../../../../core/design_system/constants/app_icons.dart';
+import '../../../../../core/design_system/constants/app_icons.dart';
+
 
 class AddressItem extends StatelessWidget {
   final String address;
