+++ lib/ui/screens/store_details/about_store/widgets/rating_bar_item.dart
