+++ lib/ui/screens/store_details/about_store/widgets/store_map_view.dart
