+++ lib/ui/screens/store_details/store_details_screen.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:sellio_mobile/core/design_system/constants/app_icons.dart';
 import 'package:sellio_mobile/core/design_system/themes/sellio_theme_provider.dart';
-import 'package:sellio_mobile/ui/screens/about_store/about_store.dart';
 import 'package:sellio_mobile/ui/screens/store_details/widgets/store_header.dart';
 import 'package:sellio_mobile/ui/screens/store_details/widgets/store_info_card.dart';
 
+import 'about_store/about_store.dart';
 import 'widgets/featured_items_section.dart';
 import 'widgets/store_category_tabs.dart';
 import 'widgets/store_products_list.dart';
